Refactor out TokenParser
